requirements.txt
asttokens==2.2.1
async-timeout==4.0.2
azure-core==1.26.2
azure-storage-blob==12.14.1
backcall==0.2.0
bleach==5.0.1
blinker==1.6.2
boto3==1.16.63
botocore==1.19.63
click==8.1.4
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.6
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-swagger==2.2.0
django-rq==2.5.1
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-generators==0.3.0
drf-yasg==1.20.0
ecdsa==0.18.0
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
Flask==2.3.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.51.1
grpcio-status==1.48.2
ipykernel==6.20.2
ipython==8.8.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
jupyter_client==7.4.9
jupyter_core==5.1.5
label-studio==1.7.0
label-studio-converter==0.0.48rc0
label-studio-tools==0.0.1
launchdarkly-server-sdk==7.5.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msrest==0.7.1
mysqlclient==2.2.0
nest-asyncio==1.5.6
nltk==3.6.7
openapi-codec==1.3.2
ordered-set==4.0.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.36
proto-plus==1.22.2
pure-eval==0.2.2
pyasn1==0.5.0
pycryptodome==3.18.0
Pygments==2.14.0
python-dateutil==2.8.2
pywin32==305
pyzmq==25.0.0
redis==4.4.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.21
s3transfer==0.3.7
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
virtualenv==20.16.5
Werkzeug==2.3.6
